job-manager: centralize out job state machine
Move the code used to replay the job eventlog
during job manager restart into event.c with the
goal of centralizing the logic that drives state
transitions with events.

In this commit it is still only used from
job_create_from_eventlog() during restart.
