+++ src/modules/job-manager/job.c
@@ -12,9 +12,10 @@
 #include "config.h"
 #endif
 #include <stdlib.h>
-#include "src/common/libjob/job.h"
+#include <flux/core.h>
+
 #include "job.h"
-#include "util.h"
+#include "event.h"
 
 void job_decref (struct job *job)
 {
@@ -50,67 +51,25 @@ struct job *job_create_from_eventlog (flux_jobid_t id, const char *s)
 {
     struct flux_kvs_eventlog *eventlog;
     const char *event;
-    double timestamp;
-    char name[FLUX_KVS_MAX_EVENT_NAME + 1];
-    char context[FLUX_KVS_MAX_EVENT_CONTEXT + 1];
-    bool submit_valid = false;
     struct job *job;
 
     if (!(job = job_create ()))
         return NULL;
     job->id = id;
-    job->state = FLUX_JOB_SCHED;
     if (!(eventlog = flux_kvs_eventlog_decode (s)))
         goto error;
     event = flux_kvs_eventlog_first (eventlog);
     while (event) {
-        if (flux_kvs_event_decode (event, &timestamp,
-                                   name, sizeof (name),
-                                   context, sizeof (context)) < 0)
+        if (event_job_update (job, event) < 0)
             goto error;
-        if (!strcmp (name, "submit")) {
-            int priority, userid, flags;
-            if (util_int_from_context (context, "priority", &priority) < 0)
-                goto error;
-            if (util_int_from_context (context, "userid", &userid) < 0)
-                goto error;
-            if (util_int_from_context (context, "flags", &flags) < 0)
-                goto error;
-            job->t_submit = timestamp;
-            job->userid = userid;
-            job->priority = priority;
-            job->flags = flags;
-            submit_valid = true;
-        }
-        else if (!strcmp (name, "priority")) {
-            int priority;
-            if (util_int_from_context (context, "priority", &priority) < 0)
-                goto error;
-            job->priority = priority;
-        }
-        else if (!strcmp (name, "exception")) {
-            int severity;
-            if (util_int_from_context (context, "severity", &severity) < 0)
-                goto error;
-            if (severity == 0)
-                job->state = FLUX_JOB_CLEANUP;
-        }
-        else if (!strcmp (name, "alloc")) {
-            job->has_resources = 1;
-            job->state = FLUX_JOB_RUN;
-        }
-        else if (!strcmp (name, "free")) {
-            job->has_resources = 0;
-            job->state = FLUX_JOB_CLEANUP;
-        }
         event = flux_kvs_eventlog_next (eventlog);
     }
-    if (!submit_valid) {
-        errno = EINVAL;
-        goto error;
-    }
+    if (job->state == FLUX_JOB_NEW)
+        goto inval;
     flux_kvs_eventlog_destroy (eventlog);
     return job;
+inval:
+    errno = EINVAL;
 error:
     job_decref (job);
     flux_kvs_eventlog_destroy (eventlog);
