+++ src/modules/job-manager/job.h
@@ -36,8 +36,6 @@ struct job *job_incref (struct job *job);
 
 struct job *job_create (void);
 
-/* (re-)create job by replaying its KVS eventlog.
- */
 struct job *job_create_from_eventlog (flux_jobid_t id, const char *eventlog);
 
 #endif /* _FLUX_JOB_MANAGER_JOB_H */
